refactor: GroupScheduleTypeTransform 함수 -> 전부 ScheduleTypeTransform 클래스 내부로 통합
